requirements.txt
boto==2.40.0
dj-database-url==0.4.1
Django==1.9.6
django-crispy-forms==1.6.0
django-storages==1.4.1
gunicorn==19.5.0
Pillow==3.2.0
psycopg2==2.6.1
requests==2.10.0
whitenoise==3.1
